+++ packages/extension/lib/extension.dart
@@ -1,3 +1,5 @@
-export 'src/date.dart';
-export 'src/string.dart';
+export 'src/date/date.dart';
+export 'src/date/time.dart';
 export 'src/iterable.dart';
+export 'src/num.dart';
+export 'src/string.dart';